wip, writing tests for file-manager, renamed some files, created test data
