--- __mocks__/ignore-fs.ts
@@ -1,4 +1,0 @@
-// const mockFs = jest.createMockFromModule("fs");
-// mockFs.
-// console.log(mockFs);
-// export default mockFs;
