--- aws/README.md
@@ -1,54 +1,0 @@
-# AWS CLI v2
-
-This bundle contains a built executable of the AWS CLI v2.
-
-## Installation
-
-To install the AWS CLI v2, run the `install` script:
-```
-$ sudo ./install 
-You can now run: /usr/local/bin/aws --version
-```
-This will install the AWS CLI v2 at `/usr/local/bin/aws`.  Assuming
-`/usr/local/bin` is on your `PATH`, you can now run:
-```
-$ aws --version
-```
-
-
-### Installing without sudo
-
-If you don't have ``sudo`` permissions or want to install the AWS
-CLI v2 only for the current user, run the `install` script with the `-b`
-and `-i` options:
-```
-$ ./install -i ~/.local/aws-cli -b ~/.local/bin
-``` 
-This will install the AWS CLI v2 in `~/.local/aws-cli` and create
-symlinks for `aws` and `aws_completer` in `~/.local/bin`. For more
-information about these options, run the `install` script with `-h`:
-```
-$ ./install -h
-```
-
-### Updating
-
-If you run the `install` script and there is a previously installed version
-of the AWS CLI v2, the script will error out. To update to the version included
-in this bundle, run the `install` script with `--update`:
-```
-$ sudo ./install --update
-```
-
-
-### Removing the installation
-
-To remove the AWS CLI v2, delete the its installation and symlinks:
-```
-$ sudo rm -rf /usr/local/aws-cli
-$ sudo rm /usr/local/bin/aws
-$ sudo rm /usr/local/bin/aws_completer
-```
-Note if you installed the AWS CLI v2 using the `-b` or `-i` options, you will
-need to remove the installation and the symlinks in the directories you
-specified.
