+++ package.json
@@ -37,6 +37,7 @@
     "@babel/preset-typescript": "^7.12.16",
     "@types/jest": "^26.0.20",
     "@types/line-reader": "^0.0.33",
+    "@types/mock-fs": "^4.13.0",
     "@types/node": "^14.14.25",
     "babel-jest": "^26.6.3",
     "jest": "^26.6.3",
@@ -48,6 +49,7 @@
   },
   "dependencies": {
     "enquirer": "^2.3.6",
-    "line-reader": "^0.4.0"
+    "line-reader": "^0.4.0",
+    "mock-fs": "^4.13.0"
   }
 }
