+++ src/file-manager/__tests__/data/two-lines.csv
@@ -1,0 +1,2 @@
+line1
+line2
