+++ src/translater/index.ts
