+++ src/view/__tests__/View.test.ts
@@ -1,5 +1,5 @@
-import Mediator from "../../Mediator/Mediator";
-import View from "../View";
+import Mediator from "../../mediator/mediator";
+import View from "../view";
 
 const view = new View();
 const mediator = new Mediator({ view });
